+++ changelog.md
@@ -4,3 +4,8 @@
 
 This release is the first release of the Optrone argument parser.
 
+# v1.1.0
+
+## Global Parameters
+
+This release introduces application-wide parameters support.